Hey there,
First of all thank you for your effort in maintaining this gem. 
We're using this gem in production and observed some weird performance issues when statsd wasn't available during the maintenance. We're still investigating what exactly happened. Anyway I played a bit with UDPSockets and did some benchmarks using this [script](https://gist.github.com/AMekss/71133597e23ec9929a34b2b6452f3e00)

Here is what it looked like:
_When statsd was available_
```
                           user     system      total        real
connect_and_send:      0.070000   0.120000   0.190000 (  0.406062)
just_send:             0.050000   0.100000   0.150000 (  0.279297)
```
_When statsd was down_
```
                           user     system      total        real
connect_and_send:      0.050000   0.100000   0.150000 (  0.156057)
just_send:             0.030000   0.050000   0.080000 (  0.081477)
```
This PR implements performance improvements based on my findings (also I fixed deprecation warnings for the spec run)