@degemer 2017-03-13
Thank you very much for your contribution @AMekss !

It's a huge performance improvement! I wanted to compare them both more precisely, so I used this [gem](https://github.com/evanphx/benchmark-ips), and slightly modified your script [there](https://gist.github.com/degemer/37958d3aa3527cc47775d1d22f23762f).

Without dogstatsd:
```bash
Warming up --------------------------------------
    connect_and_send     1.000  i/100ms
           just_send     8.000  i/100ms
Calculating -------------------------------------
    connect_and_send      4.276  (±23.4%) i/s -     21.000  in   5.043441s
           just_send     85.699  (±12.8%) i/s -    424.000  in   5.059091s

Comparison:
           just_send:       85.7 i/s
    connect_and_send:        4.3 i/s - 20.04x  slower
```

With dogstatsd:
```bash
Warming up --------------------------------------
    connect_and_send     1.000  i/100ms
           just_send     6.000  i/100ms
Calculating -------------------------------------
    connect_and_send      2.654  (± 0.0%) i/s -     14.000  in   5.321916s
           just_send     57.761  (±10.4%) i/s -    288.000  in   5.056199s

Comparison:
           just_send:       57.8 i/s
    connect_and_send:        2.7 i/s - 21.77x  slower
```

I still want to check a few potential edge cases before merging (results are too good not to have a drawback).