@AMekss 2017-03-14
Sure, its always better be safe and don't rush. Let me know if I can help