@AMekss 2017-03-14
I've noticed that I forgot to remove `@host` and `@port` from this [line](https://github.com/DataDog/dogstatsd-ruby/pull/44/files#diff-f7b44886d16ab8eb23deaed048ad0b10L430), so I just did that :)