@AMekss 2017-03-29
Recently I came across why it's faster:
> When you pass in host and port into the send method it DNS resolves before each send, which has major overhead.

[source](http://codereview.stackexchange.com/a/86996)