@bobbytables 2017-06-14
Hey @AMekss , Thanks for adding this PR to this project. For fun, I thought I'd show you our graph after shipping this.

![pasted_image_6_14_17__4_03_pm](https://user-images.githubusercontent.com/521627/27152195-10420af8-511b-11e7-99ba-9bc225de1f2d.png)